+++ pyswtools/stat/main.py
@@ -2,19 +2,28 @@
 Give stat from an assembly
 """
 
-import click
-
-"""
-Auto export of solidworks files to other extensions
-"""
-
 import os
-import math
+from enum import Enum
 import click
 
 # pylint: disable=relative-beyond-top-level
 from ..utils import check_system, check_system_verbose
 from ..config import get_config
+
+
+class TypeOutput(str, Enum):
+    """Class represeting an output type"""
+
+    TREE = "tree"
+    LIST = "list"
+
+
+class TypeSort(str, Enum):
+    """Class represeting an output type"""
+
+    MASS = "mass"
+    NAME = "name"
+
 
 if check_system():
     # pylint: disable=import-error
@@ -25,13 +34,19 @@ if check_system():
     VT_Views = win32com.client.VARIANT(pythoncom.VT_VARIANT, Views)
     VT_BYREF = win32com.client.VARIANT(pythoncom.VT_BYREF | pythoncom.VT_I4, -1)
 
-#TODO define type_sort and type_output as dataclass
 
-def sort_key_struct(struct, type_sort):
-    if type_sort == "name":
-        return sorted(struct.keys(),key=str.lower)
-    if type_sort == "mass":
-        return sorted(struct.keys(), key=lambda i: struct[i]["mass"]*struct[i]["number"],reverse=True )
+def sort_key_struct(struct: dict, type_sort: TypeSort = TypeSort.MASS) -> list:
+    """
+    Sort a dict struct following a given sort
+    """
+    if type_sort is TypeSort.NAME:
+        return sorted(struct.keys(), key=str.lower)
+    if type_sort is TypeSort.MASS:
+        return sorted(
+            struct.keys(),
+            key=lambda i: struct[i]["mass"] * struct[i]["number"],
+            reverse=True,
+        )
 
     return struct.keys()
 
@@ -40,7 +55,12 @@ def print_header():
     click.echo(header)
     click.echo("-" * len(header))
 
-def display_tree(tree_struct_mass, type_sort="name", indent=""):
+def display_tree(
+    tree_struct_mass: dict, type_sort: TypeSort = TypeSort.NAME, indent: str = ""
+) -> None:
+    """
+    Display a tree struct recursively
+    """
     sorted_list = sort_key_struct(tree_struct_mass, type_sort)
     n = len(sorted_list)
     if indent == "":
@@ -56,52 +76,81 @@ def display_tree(tree_struct_mass, type_sort="name", indent=""):
                 char= "  "
             display_tree(tree_struct_mass[elem]["children"],type_sort, indent + char)
 
-def display_list(list_struct, type_sort="name"):
+
+def display_list(list_struct: dict, type_sort: TypeSort = TypeSort.NAME) -> None:
+    """
+    Display a list struct
+    """
     print_header()
     for elem in sort_key_struct(list_struct, type_sort):
-        if abs(list_struct[elem]['density'] -1000) < 1e-4 or True:
-            click.echo(f"{'- ' + elem :<50} | {list_struct[elem]['mass'] * list_struct[elem]['number']:5.3f} | {list_struct[elem]['number']:3d} | {list_struct[elem]['mass']:6.4f} | {list_struct[elem]['density']:6.4f}")
+        click.echo(
+            f"{'- ' + elem :<50} | {list_struct[elem]['mass'] * list_struct[elem]['number']:5.3f} | {list_struct[elem]['number']:3d} | {list_struct[elem]['mass']:6.4f} | {list_struct[elem]['density']:6.4f}"
+        )
 
-def get_clean_name(sw_comp):
+
+def get_clean_name(sw_comp) -> str:
+    """
+    Get the cleaned name from a component. remove the number from the component
+    """
     return sw_comp.Name2.rpartition("-")[0].rpartition("/")[-1]
 
-def complete_info_on_list(sw_comp_children, dict_of_comp):
+
+def complete_info_on_list(sw_comp_children, dict_of_comp: dict) -> dict:
+    """
+    Get all the info from a list of component
+    """
     children = {}
     for sw_child in sw_comp_children:
+        # Get the name
         sw_child_name = get_clean_name(sw_child)
+
+        # Get the info fo the assembly
         child = complete_info_assembly(sw_child, dict_of_comp)
-        if sw_child_name not in children.keys():
+
+        # If the child is already here, only increase the number
+        if sw_child_name not in children:
             children[sw_child_name] = child
         else:
             children[sw_child_name]["number"] += 1
 
     return children
 
-def complete_info_assembly(sw_comp, dict_of_comp):
+
+def complete_info_assembly(sw_comp, dict_of_comp: dict) -> dict:
+    """
+    Get all the information from an assembly
+    """
+
+    # Get name of component
     sw_comp_name = get_clean_name(sw_comp)
 
+    # If already there only increase the number
     if sw_comp_name in dict_of_comp:
         dict_of_comp[sw_comp_name]["number"] += 1
     else:
+        # Otherwise get the mass
         sw_comp_doc_ext = sw_comp.GetModelDoc2.Extension
         sw_mass = sw_comp_doc_ext.CreateMassProperty2.Mass
 
+        # Create an new entity in the general dict
         dict_of_comp[sw_comp_name] = {
-            "mass": sw_mass,
-            "density": sw_comp_doc_ext.CreateMassProperty2.density,
             "number": 1,
+            "mass": sw_mass,
+            "density": sw_comp_doc_ext.CreateMassProperty2.density
         }
-    
+
+    # Get info about children
     sw_comp_children = sw_comp.GetChildren
     children = complete_info_on_list(sw_comp_children, dict_of_comp)
 
-
+    # Create the entry in the tree dict
     return {
         "mass": dict_of_comp[sw_comp_name]["mass"],
         "density": dict_of_comp[sw_comp_name]["density"],
         "number": 1,
         "children": children,
     }
+
 
 @click.command()
 @click.help_option("-h", "--help")
@@ -110,13 +159,12 @@ def complete_info_assembly(sw_comp, dict_of_comp):
     type=click.Path(
         exists=True,
         dir_okay=False,
-        ),
+    ),
 )
-@click.option('--tree', 'type_output', flag_value='tree', default=True)
-@click.option('--list', 'type_output', flag_value='list')
-@click.option('--mass', 'type_sort', flag_value='mass', default=True)
-@click.option('--name', 'type_sort', flag_value='name')
-def stat(input_path, type_output, type_sort) -> None:
+
+@click.option("--type_output", "type_output", type=click.Choice(TypeOutput), default=TypeOutput.TREE)
+@click.option("--type_sort", "type_sort", type=click.Choice(TypeSort), default=TypeSort.MASS)
+def stat(input_path: str, type_output: TypeOutput, type_sort: TypeSort) -> None:
     """
     Display stat about an assembly
     """
@@ -128,7 +176,7 @@ def stat(input_path, type_output, type_sort) -> None:
         click.echo(f"{input_path} is not an assembly file")
         return
 
-    path_directory, filename = os.path.split(input_path)
+    _, filename = os.path.split(input_path)
 
     conf = get_config()
 
@@ -142,9 +190,10 @@ def stat(input_path, type_output, type_sort) -> None:
     # Activate it
     sw_app.ActivateDoc3(filename, True, 2, VT_BYREF)
 
+    # Get list of components
     sw_comps = sw_doc.GetComponents(True)
- 
 
+    # Empty struct to fill
     assembly_name = filename.rpartition('.')[-0]
     dict_of_comp = {
         f"{assembly_name}" : {
@@ -163,9 +212,9 @@ def stat(input_path, type_output, type_sort) -> None:
         }
     }
 
-    if type_output == "tree":
+    if type_output is TypeOutput.TREE:
         display_tree(tree_of_comp, type_sort)
-    elif type_output == "list":
+    elif type_output is TypeOutput.LIST:
         display_list(dict_of_comp, type_sort)
     else:
-        pass
+        click.echo(f"Type output {type_output} is unknown {type_output is TypeOutput.TREE} {type(type_output)}")